+++ dll/win32/imm32/CMakeLists.txt
@@ -8,7 +8,13 @@ add_definitions(-D_WIN32_WINNT=0x600)
 spec2def(imm32.dll imm32.spec ADD_IMPORTLIB)
 
 list(APPEND SOURCE
-    imm.c
+    candidate.c
+    guideline.c
+    ime.c
+    imm32.c
+    imp.c
+    keymsg.c
+    utils.c
     ${CMAKE_CURRENT_BINARY_DIR}/imm32_stubs.c
     ${CMAKE_CURRENT_BINARY_DIR}/imm32.def)
